titanic-serverless-ml-intro /requirements.txt
hopsworks
joblib
scikit-learn
seaborn
dataframe-image
modal-client
gradio
numpy